combine password and user input
