+++ app/components/common/userInput/index.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import UserInput from '.';
 
 it('renders a input field', () => {
@@ -34,4 +34,20 @@ it('checks input value to be empty', () => {
   render(<UserInput label="Username" placeholder="enter your name" value="" setValue={() => {}} />);
   const inputElement = screen.getByTestId('user-input');
   expect(inputElement).toHaveValue('');
+});
+
+it('checks show password', () => {
+  render(
+    <UserInput
+      type="password"
+      label="New Password"
+      description="Password must be 8 characters"
+      value=""
+      setValue={() => {}}
+    />
+  );
+  const passwordInput = screen.getByTestId('user-input');
+  const eyeIcon = screen.getByTestId('icon');
+  fireEvent.click(eyeIcon);
+  expect(passwordInput).toHaveAttribute('type', 'text');
 });
